pyproject.toml
[tool.poetry]
name = "pythonproject"
version = "0.1.0"
description = ""
authors = ["Artsiom_PH <workartpharm@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
scrapy = "^2.11.0"
psycopg2-binary = "^2.9.9"
httpx = "^0.26.0"


[tool.poetry.group.dev.dependencies]
black = "^23.12.1"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
